requirements.txt
absl-py==2.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
astunparse==1.6.3
async-timeout==5.0.1
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2025.6
cryptography==46.0.3
cssbeautifier==1.15.4
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.8
django-crispy-forms==2.5
mysqlclient==2.2.6
django-csp==4.0
django-dotenv==1.4.2
django-filter==25.1
django-otp==1.6.3
django-ratelimit==4.1.0
django-redis==6.0.0
djlint==1.36.4
dnspython==2.8.0
EditorConfig==0.17.1
et_xmlfile==2.0.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
joblib==1.5.2
jsbeautifier==1.15.4
json5==0.12.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.4
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
patsy==1.0.2
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-magic==0.4.27
python-whois==0.9.6
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.0.1
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
ruff==0.5.7
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.5
structlog==25.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.11.0
wrapt==2.0.1
